requirements.txt
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
attrs==24.2.0
awsebcli==3.20.10
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
boto3==1.35.0
botocore==1.35.0
build==1.2.1
cachetools==5.4.0
cement==2.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cohere==5.8.1
colorama==0.4.3
coloredlogs==15.0.1
crewai==0.51.1
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
Django==5.1
djangorestframework==3.15.2
docstring_parser==0.16
embedchain==0.1.121
fastapi==0.112.1
fastavro==1.9.5
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.62.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
gptcache==0.1.44
grpc-google-iam-v1==0.13.1
grpcio==1.65.5
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.5
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.2
instructor==1.3.3
jiter==0.4.2
jmespath==1.0.1
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
kubernetes==30.1.0
langchain==0.2.14
langchain-cohere==0.1.9
langchain-community==0.2.12
langchain-core==0.2.33
langchain-experimental==0.0.64
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langsmith==0.1.99
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mem0ai==0.0.20
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.0
openai==1.41.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pathspec==0.10.1
portalocker==2.10.1
posthog==3.5.0
proto-plus==1.24.0
protobuf==4.25.4
pulsar-client==3.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
qdrant-client==1.11.0
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
schema==0.7.7
semantic-version==2.8.5
setuptools==72.2.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
sqlparse==0.5.1
starlette==0.38.2
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
termcolor==1.1.0
tiktoken==0.7.0
tokenizers==0.20.0
tqdm==4.66.5
typer==0.12.4
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.19
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
wcwidth==0.1.9
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.20.0
